added proxy trait
